+++ src/theme/elements/Checkbox.js
@@ -1,23 +1,23 @@
-import { css, keyframes } from 'styled-components'
+import { css } from 'styled-components'
 import { th } from '@pubsweet/ui-toolkit'
 
-const checking = keyframes`
-  0% {
-    transform: scale(0.8);
-  }
+// const checking = keyframes`
+//   0% {
+//     transform: scale(0.8);
+//   }
 
-  20% {
-    transform: scale(1.2);
-  }
+//   20% {
+//     transform: scale(1.2);
+//   }
 
-  80% {
-    transform: scale(1);
-  }
+//   80% {
+//     transform: scale(1);
+//   }
 
-  100% {
-    transform: scale(1);
-  }
-`
+//   100% {
+//     transform: scale(1);
+//   }
+// `
 
 const localBorderSize = '3px'
 const localBorderTwoSize = '1px'
@@ -59,7 +59,6 @@ export default {
       color: ${th('colorPrimary')};
 
       &:before {
-        animation: ${checking} 0.5s;
         box-shadow: 0 0 0 ${localBorderTwoSize} ${th('colorPrimary')};
       }
     }