+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### [0.0.3](https://gitlab.coko.foundation///compare/v0.0.2...v0.0.3) (2020-11-01)
 
-
 ### Features
 
-* **theme:** export theme ([87837ba](https://gitlab.coko.foundation///commit/87837badf3d4df3e25b683dee89e92db2c3fbec3))
+- **theme:** export theme ([87837ba](https://gitlab.coko.foundation///commit/87837badf3d4df3e25b683dee89e92db2c3fbec3))
 
 ### 0.0.2 (2020-11-01)